split up main restify.js into multiple files for clarity
